Attempting to migrate admin route to it's own blueprint.
I've made so many changes everywhere, I've probably broken something.
If only I had a test suite.
