--- undyingkingdoms/admin/templates/admin.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Admin Panel</title>
-</head>
-<body>
-	<p>Admin page that needs a fancy template panel for all these nice functions.</p>
-	<li>
-    <ol><a href="/">Home</a></ol>
-    <ol><button onclick="callReset();">Reset</button></ol>
-    <ol><button value="/admin/refresh" onclick="callRefresh();">Refresh</button></ol>
-	</li>
-	<div>
-    <p>JSON response:</p>
-    <pre id="response">Push a button ...</pre>
-	</div>
-	<script type="text/javascript" src="{{ url_for('admin.static', filename='admin.js') }}"></script>
-</body>
-</html>
