+++ undyingkingdoms/blueprints/admin/metadata.py
@@ -1,0 +1,4 @@
+bot_county_prefix = ["Eldritch", "Dengmar", "Zort", "Pralde", "Eggroth", "Wuanlo"]
+bot_county_suffix = ["ville", " Town", "ion", "oth", " Land"]
+bot_leader_prefix = ["Skri", "Aldar", "Frel", "Eldr", "Toth", "Zon", "Pandish", "Gresk", "Linsk"]
+bot_leader_suffix = ["oth", "ith", "ion", "ilisk", "anth", "ills", "ondo", "alasl"]
