+++ undyingkingdoms/templates/mobile/gameplay/layout.html
@@ -28,20 +28,21 @@ nav > a {
 {% set county = current_user.county %}
 {% set kingdom = county.kingdom %}
 <nav class="mobile-sidebar">
-	<a href="{{ url_for('overview', kingdom_id=0, county_id=0) }}">County Overview</a>
+	<a href="{{ url_for('overview', kingdom_id=0, county_id=0) }}">County&nbsp;Overview</a>
 	<a href="{{ url_for('economy') }}">Economy</a>
 	<a href="{{ url_for('infrastructure') }}">Infrastructure</a>
 	<a href="{{ url_for('military') }}">Military</a>
 	<a href="{{ url_for('infiltration') }}">Infiltration</a>
 	<a href="{{ url_for('messages') }}" {% if has_mail %}style="color:blue;font-weight:bold;"{% endif %}>Inbox</a>
-	<a href="{{ url_for('chatroom_api') }}">Town Hall</a>
-	<a href="{{ url_for('kingdom', kingdom_id=kingdom.id) }}">Kingdom Overview</a>
+	<a href="{{ url_for('chatroom_api') }}">Town&nbsp;Hall</a>
+	<a href="{{ url_for('kingdom', kingdom_id=kingdom.id) }}">Kingdom&nbsp;Overview</a>
 	<br>
 	<a href="{{ url_for('achievements') }}">Achievements</a>
 	<a href="{{ url_for('forum', thread_id=0, post_id=0) }}">Forum</a>
-	<a href="{{ url_for('guide') }}">New Player Guide</a>
+	<a href="{{ url_for('guide') }}">New&nbsp;Player&nbsp;Guide</a>
 	<a href="{{ url_for('leaderboard') }}">Leaderboard</a>
-	<a href="{{ url_for('versions') }}">Read About Updates</a>
+	<a href="{{ url_for('versions') }}">Read&nbsp;About&nbsp;Updates</a>
+	{% if user.is_admin %}<a href="{{ url_for('admin.home_api') }}">Admin</a>{% endif %}
 	<a href="{{ url_for('logout') }}">Logout</a>
 </nav>
 {% endblock %}
