+++ undyingkingdoms/templates/mobile/user/admin.html
@@ -1,0 +1,21 @@
+{% extends "mobile/gameplay/layout.html" %}
+
+{% block title %}Admin: {{ super() }}{% endblock %}
+
+{% block style %}
+{{ super() }}
+<style type="text/css">
+
+</style>
+{% endblock %}
+
+{% block content1 %}
+<div id="content">
+  <form method="POST" accept-charset="UTF-8" role="form">
+    {{ form.csrf_token }}
+    {{ form.tax }}
+    <button type="submit">Add 3 bot counties</button>
+    <br>
+  </form>
+</div>
+{% endblock %}
